支持通过环境变量配置自动更新

+++ docker/run.sh
@@ -1,16 +1,17 @@
 #!/bin/bash
 
-port="7789"
+update="$AUTO_UPDATE"
 path="./"
 jar="ani-rss-jar-with-dependencies.jar"
 jar_path=$path$jar
 
-if [ ! -f $jar_path ]; then
+if [ ! -f "$jar_path" ] && [ "$update" ]; then
+    echo "正在更新……"
     version=$(curl -Ls https://github.com/wushuo894/ani-rss/releases/latest | grep -o '<h1 data-view-component="true" class="d-inline mr-3">[^<]*</h1>' | sed 's/<[^>]*>//g')
     url="https://github.com/wushuo894/ani-rss/releases/download/$version/ani-rss-jar-with-dependencies.jar"
-    wget -O $jar_path $url
 
-    if [ $? -eq 0 ]; then
+    if ! wget -O "$jar_path" "$url"
+    then
         echo "$jar_path 下载成功！"
     else
         echo "$jar_path 下载失败。"
@@ -26,8 +27,8 @@ fi
 
 while :
 do
-    java -jar -Xmx2g $jar_path --port $port
-    if [ $? -ne 0 ]; then
+    if ! java -jar -Xmx2g $jar_path --port "$PORT"
+    then
         break
     fi
 done
